Added object transfer for server (#46)
* Added Server.java

* Added Client.java
Added auto-connect between client and server
Close #39

* Formatted help text

* Replaced with new task for lab 6

* fix port number input

* Added implementation of Serializable and serialVersionUID  for classes:
- Color
- Coordinates
- Movie
- Person

* Added saveObject method for CollectionManager

* Added add command implementation to client's side with sending serialized object to server

* Added auto saving of collection after terminating session

* Commands were split to hasInput
Commands input {String... args} -> {Object inputData}

* Added object transfer for server

---------

+++ src/main/java/classes/movie/Coordinates.java
@@ -18,10 +18,9 @@ public class Coordinates implements Comparable<Coordinates>, Serializable {
     private long x; //Максимальное значение поля: 279
     private int y; //Значение поля должно быть больше -230
 
-    private Coordinates() {
-    }
+    private Coordinates() {}
 
-    public Coordinates(long x, int y) throws NotGreatThanException, NullValueException, GreatThanException {
+    public Coordinates(long x, long y) throws NotGreatThanException, NullValueException, GreatThanException {
         this.x = new FieldHandler(x, MAX_VALUE).handleLong();
         this.y = new FieldHandler(y, GREAT_THAN_X).handleInt();
     }
